+++ src/providers/ModalProvider/modal-reducer.js
@@ -19,7 +19,8 @@ export const initialState = {
     withButtons: false,
     components: []
   },
-  containerStyle: {}
+  containerStyle: {},
+  exitable: false
 };
 
 export const reducer = (state, action) => {